requirements.txt
pytest==7.4.4
Werkzeug==2.3.6
python-dotenv==1.0.0
mysql-connector-python==8.2.0